fix clang-tidy

+++ tasks/vlasova_a_image_smoothing/mpi/src/ops_mpi.cpp
@@ -34,7 +34,7 @@ std::uint8_t ComputePixelMedian(int col_idx, int row_idx, int overlap_start,
   }
 
   if (!neighbors.empty()) {
-    auto middle = neighbors.begin() + neighbors.size() / 2;
+    auto middle = neighbors.begin() + static_cast<std::size_t>(neighbors.size()) / 2;
     std::nth_element(neighbors.begin(), middle, neighbors.end());
     return *middle;
   }
